@@ -42,6 +42,7 @@
     "autoprefixer": "^6.5.1",
     "babel-core": "^6.18.2",
     "babel-loader": "^6.2.5",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-plugin-transform-object-rest-spread": "^6.5.0",
     "babel-plugin-transform-runtime": "^6.7.5",
     "babel-preset-es2015": "^6.3.13",
@@ -94,7 +95,7 @@
     "serialize-javascript": "^1.3.0",
     "slug": "^0.9.1",
     "uid-safe": "^2.0.0",
-    "webpack": "^2.3.2",
+    "webpack": "^2.6.1",
     "whatwg-fetch": "^1.0.0"
   },
   "devDependencies": {
