@@ -1,26 +1,27 @@
+import React from 'react';
 import { BlogIndexPage, PostPage } from './components';
 import { loadPost, loadPosts } from 'posts/actions';
 import { getPost, getPosts } from 'posts/selectors';
 
 export default {
-  index(req, res, getState) {
-    return res.dispatch(loadPosts())
-      .then(() => res.render(
-        BlogIndexPage,
-        {posts: getPosts(getState())},
-        {title: 'Blog'}
-      ));
+  index(req, store) {
+    return store.dispatch(loadPosts())
+      .then(() => {
+        const posts = getPosts(store.getState());
+        return <BlogIndexPage posts={posts} />;
+      });
   },
 
-  view(req, res, getState) {
-    let permalink = req.params.post;
+  view(req, store) {
+    const permalink = req.params.post;
 
-    return res.dispatch(loadPost(permalink))
+    return store.dispatch(loadPost(permalink))
       .then(() => {
-        let post = getPost(getState(), permalink);
+        const post = getPost(store.getState(), permalink);
         if (post) {
-          res.render(PostPage, {post}, {title: post.title});
+          return <PostPage post={post} />;
         } else {
+          // TODO: Fix
           res.render404();
         }
       });
