@@ -1,3 +1,4 @@
+import React from 'react';
 import { HomePage } from './components';
 import { loadHomePage } from './actions';
 import { getPage } from 'pages/selectors';
@@ -5,14 +6,13 @@ import { loadPosts } from 'posts/actions';
 import { getPosts } from 'posts/selectors';
 
 export default {
-  index(req, res, getState) {
-    return Promise.all([res.dispatch(loadHomePage()), res.dispatch(loadPosts())])
+  index(req, store) {
+    return Promise.all([store.dispatch(loadHomePage()), store.dispatch(loadPosts())])
       .then(() => {
-        let state = getState();
-        res.render(HomePage, {
-          content: getPage(state, 'home').content,
-          posts: getPosts(state)
-        });
+        const state = store.getState();
+        const content = getPage(state, 'home').content;
+        const posts = getPosts(state);
+        return <HomePage content={content} posts={posts} />;
       });
   }
 };
