@@ -1,14 +1,14 @@
+import React from 'react';
 import { ProjectsPage } from './components';
 import { loadProjectsPage } from './actions';
 import { getPage } from 'pages/selectors';
 
 export default {
-  index(req, res, getState) {
-    return res.dispatch(loadProjectsPage())
-      .then(() => res.render(
-        ProjectsPage,
-        {categories: getPage(getState(), 'projects').categories},
-        {title: 'Projects'}
-      ));
+  index(req, store) {
+    return store.dispatch(loadProjectsPage())
+      .then(() => {
+        const { categories } = getPage(store.getState(), 'projects');
+        return <ProjectsPage categories={categories} />;
+      });
   }
 };
